Class that contains all isomiRs annotation for all samples

The IsomirDataSeq is a subclass of SummarizedExperiment. used to store the raw data, intermediate calculations and results of an miRNA/isomiR analysis. This class stores all raw isomiRs data for each sample, processed information, summary for each isomiR type, raw counts, normalized counts, and table with experimental information for each sample.

Details

IsomirDataSeqFromFiles creates this object using seqbuster output files.

Methods for this objects are counts() to get count matrix and isoSelect() for miRNA/isomiR selection. Functions available for this object are isoCounts() for count matrix creation, isoNorm() for normalization, isoDE() for differential expression and isoPLSDA() for clustering. isoPlot() helps with basic expression plot.

metadata contains one list:

rawData is a data.frame with the information of each sequence found in the data and the counts for each sample.

The naming of isomiRs follows these rules:

miRNA name

type:ref if the sequence is the same than the miRNA reference. iso if the sequence has variations.

iso_5p tag:indicates variations at 5 position. The naming contains two words: direction - nucleotides, where direction can be UPPER CASE NT (changes upstream of the 5 reference position) or LOWER CASE NT (changes downstream of the 5 reference position). 0 indicates no variation, meaning the 5 position is the same than the reference. After direction, it follows the nucleotide/s that are added (for upstream changes) or deleted (for downstream changes).

iso_3p tag:indicates variations at 3 position. The naming contains two words: direction - nucleotides, where direction can be LOWER CASE NT (upstream of the 3 reference position) or UPPER CASE NT (downstream of the 3 reference position). 0 indicates no variation, meaning the 3 position is the same than the reference. After direction, it follows the nucleotide/s that are added (for downstream changes) or deleted (for upstream chanes).

iso_add tag:indicates nucleotides additions at 3 position. The naming contains two words: direction - nucleotides, where direction is UPPER CASE NT (upstream of the 5 reference position). 0 indicates no variation, meaning the 3 position has no additions. After direction, it follows the nucleotide/s that are added.

iso_snv tag: indicates nucleotides substitutions along the sequences. The naming contains three words: position-nucleotide@isomiR-nucleotide@reference.

iso_snv_seed tag: same than iso_snv tag, but only if the change happens between nucleotide 2 and 8.

In general nucleotides in UPPER case mean insertions respect to the reference sequence, and nucleotides in LOWER case mean deletions respect to the reference sequence.

Examples

path <- system.file("extra", package="isomiRs")
fn_list <- list.files(path, pattern="mirna", full.names = TRUE)
de <- data.frame(row.names=c("f1" , "f2"),
                 condition = c("newborn", "newborn"))
ids <- IsomirDataSeqFromFiles(fn_list, coldata=de)

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 1.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 2.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 4.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 5.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 6.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 13.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 14.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 15.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 16.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 17.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 18.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 19.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 20.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 21.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 22.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 23.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 24.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 25.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 26.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 27.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 28.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 29.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 36.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 37.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 40.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 41.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 42.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 43.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 44.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 45.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 46.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 49.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 51.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 52.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 53.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 61.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 62.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 63.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 67.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 68.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 69.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 70.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 71.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 72.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 73.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 76.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 81.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 91.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 92.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 93.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 94.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 95.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 96.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 98.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 99.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 100.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 101.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 102.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 103.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 106.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 107.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 108.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 109.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 110.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 112.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 113.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 114.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 115.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 116.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 117.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 118.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 119.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 120.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 121.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 122.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 123.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 124.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 125.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 126.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 128.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 129.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 131.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 132.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 133.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 134.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 135.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 136.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 137.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 138.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 139.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 140.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 146.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 147.

#> Warning: `funs()` is deprecated as of dplyr 0.8.0.
#> Please use a list of either functions or lambdas: 
#> 
#>   # Simple named list: 
#>   list(mean = mean, median = median)
#> 
#>   # Auto named with `tibble::lst()`: 
#>   tibble::lst(mean, median)
#> 
#>   # Using lambdas
#>   list(~ mean(., trim = .2), ~ median(., na.rm = TRUE))
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> Total samples filtered due to low number of hits: 0

head(counts(ids))

#>                f1 f2
#> hsa-let-7a-5p   0  3
#> hsa-let-7b-5p   0 22
#> hsa-let-7d-3p  62  0
#> hsa-let-7d-5p   6  0
#> hsa-let-7e-5p   3  0
#> hsa-let-7f-5p 151  0