Import mirtop output into IsomirDataSeq

The tabular output of mirtop is compatible with IsomirDataSeq. This function allows to import the data and filter low confidence isomiRs for downstream analysis.

IsomirDataSeqFromMirtop(mirtop, coldata, ...)

Arguments

mirtop
data.frame with the output of mirtop export

coldata
data.frame with the metadata of the samples

...
It supports the same parameters as in IsomirDataSeqFromRawData.

Value

IsomirDataSeq class object.

Details

The output is generated with mirtop export --format isomir.

Examples

library(readr)
path <- system.file("extra", "mirtop", package="isomiRs")
fn <- list.files(path, full.names = TRUE)
de <- data.frame(row.names=c("sample1" , "sample2"),
                 condition = c("cc", "cc"))
# mirtop export --format isomir ....
IsomirDataSeqFromMirtop(read_tsv(fn), de)

#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   seq = col_character(),
#>   mir = col_character(),
#>   mism = col_character(),
#>   add = col_character(),
#>   t5 = col_character(),
#>   t3 = col_character(),
#>   sample1 = col_double(),
#>   sample2 = col_double()
#> )

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 1.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 6.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 7.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 8.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 11.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 12.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 13.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 18.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 19.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 20.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 21.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 22.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 23.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 24.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 25.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 26.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 27.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 28.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 29.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 30.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 31.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 32.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 33.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 34.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 39.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 40.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 41.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 49.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 50.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 54.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 55.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 56.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 57.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 58.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 59.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 60.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 61.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 62.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 63.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 67.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 69.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 70.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 89.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 90.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 91.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 92.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 93.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 94.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 95.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 96.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 101.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 104.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 113.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 114.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 115.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 116.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 117.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 119.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 120.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 121.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 122.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 123.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 128.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 129.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 130.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 131.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 132.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 133.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 135.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 136.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 137.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 138.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 139.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 140.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 141.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 142.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 143.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 144.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 145.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 147.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 149.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 150.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 151.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 152.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 153.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 154.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 155.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 156.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 157.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 162.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 163.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 164.

#> class: IsomirDataSeq 
#> dim: 66 2 
#> metadata(2): sequences rawData
#> assays(1): counts
#> rownames(66): hsa-let-7a-3p hsa-let-7a-5p ... hsa-miR-93-5p
#>   hsa-miR-98-5p
#> rowData names(0):
#> colnames(2): sample1 sample2
#> colData names(1): condition