LQNO distribution

Linear quadratic family that assumes the following relation for the variance of the normal distribution Var = mu*(1+s*mu). regression on mu and on the sigma (log and identity links)

dLQNO(x, mu = 1, sigma = 1, log = FALSE)
pLQNO(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qLQNO(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rLQNO(n, mu = 1, sigma = 1)
LQNO(mu.link="log", sigma.link="log")

dLQNO(x, mu = 1, sigma = 1, log = FALSE)

pLQNO(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)

qLQNO(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)

rLQNO(n, mu = 1, sigma = 1)

Arguments

mu.link
Type of transformation

sigma.link
Type of transformation

x
Vector of quantiles.

mu
Vector of means.

sigma
Vector of standard deviations.

log
Logical; if TRUE, probabilities p are given as log(p).

q
Vector of quantiles.

lower.tail
Logical; if TRUE (default), probabilities are P(X &lt; x) otherwise, P(X &gt; x).

log.p
Logical; if TRUE, probabilities p are given as log(p).

p
Vector of probabilities.

n
Number of observations. If length(n) > 1, the length is taken to be the number required.

Value

LQNO function

Details

Methods adapted from:

Argyropoulos, Christos, et al. "Modeling bias and variation in the stochastic processes of small RNA sequencing." Nucleic Acids Research (2017).