Accessors for the count matrix of a IsomirDataSeq object.

The counts slot holds the count data as a matrix of non-negative integer count values, one row for each isomiR, and one column for each sample. The normalized matrix can be obtained by using the parameter norm=TRUE.

counts.IsomirDataSeq(object, norm = FALSE)

# S4 method for IsomirDataSeq
counts(object, norm = FALSE)

# S4 method for IsomirDataSeq,matrix
counts(object) <- value

Arguments

object
A IsomirDataSeq object.

norm
Boolean, return log2-normalized counts.

value
An integer matrix.

Value

base::matrix with raw or normalized count data.

Examples

data(mirData)
head(counts(mirData))

#>                    pc2    pt2    pt7    pc1    pt6    pc3    pt3    pt5    pt4
#> hsa-let-7a-2-3p     11      7     10     13      4     13      9      3      0
#> hsa-let-7a-3p      928    745   1159   1293    613    973   1361    433    978
#> hsa-let-7a-5p   355578 324134 517950 507046 299028 375836 500423 152191 419754
#> hsa-let-7b-3p     1971   1410   1595   1646   1055   1267   1997    566   1148
#> hsa-let-7b-5p    77274  65928  92828 114643  53345  78586  96965  28974  71768
#> hsa-let-7c-3p       26     20     76     68     49     53     39     21     52
#>                    pc5    pc4    pc7    pc6    pt1
#> hsa-let-7a-2-3p     14     20      6     10      2
#> hsa-let-7a-3p     1614   1050   1219    637    542
#> hsa-let-7a-5p   468792 489195 340782 215635 150421
#> hsa-let-7b-3p     2852   1986   1724    875    760
#> hsa-let-7b-5p    93764  97902  68304  43050  29572
#> hsa-let-7c-3p       45     54     56     27     22