Find miRNAs target using mRNA/miRNA expression

This function creates a matrix with rows (genes) and columns (mirnas) with values indicating if miRNA-gene pair is target according putative targets and negative correlation of the expression of both molecules.

findTargets(mirna_rse, gene_rse, target, summarize = "group",
  min_cor = -0.6)

Arguments

mirna_rse
SummarizedExperiment::SummarizedExperiment with miRNA information. See details.

gene_rse
SummarizedExperiment::SummarizedExperiment with gene information. See details.

target
Matrix with miRNAs (columns) and genes (rows) target prediction values (1 if it is a target, 0 if not).

summarize
Character column name in colData(rse) to use to group samples and compare betweem miRNA/gene expression.

min_cor
Numeric cutoff for correlation value that will be use to consider a miRNA-gene pair as valid.

Value

mirna-gene matrix

Examples

data(isoExample)
mirna_ma <- matrix(rbinom(20*25, c(0, 1), 1), ncol = 20)
colnames(mirna_ma) <- rownames(mirna_ex_rse)
rownames(mirna_ma) <- rownames(gene_ex_rse)
corMat <- findTargets(mirna_ex_rse, gene_ex_rse, mirna_ma)

#> Number of mirnas 20

#> Number of genes 25

#> Factors genescontrolday1day2day3day7day14

#> Factors mirnascontrolday1day2day3day7day14

#> Order genescontrolcontrolcontrolday1day1day1day2day2day2day3day3day3day7day7day7day14day14day14

#> Order mirnascontrolcontrolcontrolday1day1day1day2day2day2day3day3day3day7day7day7day14day14day14

#> Calculating correlation matrix

#> Dimmension of cor matrix: 20 25