Find miRNAs target using mRNA/miRNA expression

This function creates a matrix with rows (genes) and columns (mirnas) with values indicating if miRNA-gene pair is target according putative targets and negative correlation of the expression of both molecules.

find_targets(mirna_rse, gene_rse, target, summarize = "group",
  min_cor = -0.6)

Arguments

mirna_rse
SummarizedExperiment::SummarizedExperiment with miRNA information. See details.

gene_rse
SummarizedExperiment::SummarizedExperiment with gene information. See details.

target
Matrix with miRNAs (columns) and genes (rows) target prediction values (1 if it is a target, 0 if not).

summarize
Character column name in colData(rse) to use to group samples and compare betweem miRNA/gene expression.

min_cor
Numeric cutoff for correlation value that will be use to consider a miRNA-gene pair as valid.

Value

mirna-gene matrix

Examples

pairs <- as.matrix(data.frame(row.names=c("gene1", "gene2"),
                            mirna1=c(0,1), mirna2=c(1,0)))
mirna_matrix <- as.matrix(data.frame(row.names=c("mirna1", "mirna2"),
                                    time0_1=c(1,1),time0_2=c(1.2,0.9),
                                    time1_1=c(8,8),time1_2=c(8.2,7.9)))
gene_matrix <- as.matrix(data.frame(row.names=c("gene1", "gene2"),
                                    time0_1=c(8,8),time0_2=c(8.2,7.9),
                                    time1_1=c(1,1),time1_2=c(1.2,0.9)))
mirna_col <- data.frame(row.names=c("time0_1","time0_2","time1_1","time1_2"),
                       group=c("t0","t0","t1","t1"))
gene_col <- data.frame(row.names=c("time0_1","time0_2","time1_1","time1_2"),
                       group=c("t0","t0","t1","t1"))

mirna <- SummarizedExperiment(assays=SimpleList(norm=as.matrix(mirna_matrix)),
                            colData= mirna_col)
gene <- SummarizedExperiment(assays=SimpleList(norm=as.matrix(gene_matrix)),
                             colData= gene_col)
find_targets(mirna, gene, pairs)

#> Number of mirnas 2

#> Number of genes 2

#> Factors genest0t1

#> Factors mirnast0t1

#> Order genest0t0t1t1

#> Order mirnast0t0t1t1

#> Calculating cor matrix

#> Dimmension of cor matrix: 2 2 

#>       mirna1 mirna2
#> gene1      0     -1
#> gene2     -1      0