Create count matrix with different summarizing options

This function collapses isomiRs into different groups. It is a similar concept than how to work with gene isoforms. With this function, different changes can be put together into a single miRNA variant. For instance all sequences with variants at 3' end can be considered as different elements in the table or analysis having the following naming hsa-miR-124a-5p.iso.t3:AAA.

isoCounts(ids, ref = FALSE, iso5 = FALSE, iso3 = FALSE, add = FALSE,
  subs = FALSE, seed = FALSE, minc = 1, mins = 1)

Arguments

ids
Object of class IsomirDataSeq.

ref
Differentiate reference miRNA from rest.

iso5
Differentiate trimming at 5 miRNA from rest.

iso3
Differentiate trimming at 3 miRNA from rest.

add
Differentiate additions miRNA from rest.

subs
Differentiate nt substitution miRNA from rest.

seed
Differentiate changes in 2-7 nts from rest.

minc
Int minimum number of isomiR sequences to be included.

mins
Int minimum number of samples with number of sequences bigger than minc counts.

Value

IsomirDataSeq object with new count table. The count matrix can be access with counts(ids).

Details

You can merge all isomiRs into miRNAs by calling the function only with the first parameter isoCounts(ids). You can get a table with isomiRs altogether and the reference miRBase sequences by calling the function with ref=TRUE. You can get a table with 5' trimming isomiRS, miRBase reference and the rest by calling with isoCounts(ids, ref=TRUE, iso5=TRUE). If you set up all parameters to TRUE, you will get a table for each different sequence mapping to a miRNA (i.e. all isomiRs).

Examples for the naming used for the isomiRs are at http://seqcluster.readthedocs.org/mirna_annotation.html#mirna-annotation.

Examples

data(mirData)
ids <- isoCounts(mirData, ref=TRUE)
head(counts(ids))

#>                        pc2    pt2    pt7    pc1    pt6    pc3    pt3    pt5
#> hsa-let-7a-2-3p.iso      6      3      0      9      2      4      4      0
#> hsa-let-7a-2-3p.ref      5      4     10      4      2      9      5      3
#> hsa-let-7a-3p.iso      782    634   1002   1140    536    857   1169    380
#> hsa-let-7a-3p.ref      146    111    157    153     77    116    192     53
#> hsa-let-7a-5p.iso    96391  84487 126832 124343  76467  96519 136940  41708
#> hsa-let-7a-5p.ref   259187 239647 391118 382703 222561 279317 363483 110483
#>                        pt4    pc5    pc4    pc7    pc6    pt1
#> hsa-let-7a-2-3p.iso      0      7     10      4      5      0
#> hsa-let-7a-2-3p.ref      0      7     10      2      5      2
#> hsa-let-7a-3p.iso      888   1313    903   1050    516    444
#> hsa-let-7a-3p.ref       90    301    147    169    121     98
#> hsa-let-7a-5p.iso    98125 130896 136026  93118  58277  39226
#> hsa-let-7a-5p.ref   321629 337896 353169 247664 157358 111195

# taking into account isomiRs and reference sequence.
ids <- isoCounts(mirData, ref=TRUE, minc=10, mins=6)
head(counts(ids))

#>                      pc2    pt2    pt7    pc1    pt6    pc3    pt3    pt5
#> hsa-let-7a-3p.iso    782    634   1002   1140    536    857   1169    380
#> hsa-let-7a-3p.ref    146    111    157    153     77    116    192     53
#> hsa-let-7a-5p.iso  96391  84487 126832 124343  76467  96519 136940  41708
#> hsa-let-7a-5p.ref 259187 239647 391118 382703 222561 279317 363483 110483
#> hsa-let-7b-3p.iso   1966   1404   1589   1639   1051   1264   1992    566
#> hsa-let-7b-5p.iso  41690  31802  41571  50956  23576  35542  48289  13555
#>                      pt4    pc5    pc4    pc7    pc6    pt1
#> hsa-let-7a-3p.iso    888   1313    903   1050    516    444
#> hsa-let-7a-3p.ref     90    301    147    169    121     98
#> hsa-let-7a-5p.iso  98125 130896 136026  93118  58277  39226
#> hsa-let-7a-5p.ref 321629 337896 353169 247664 157358 111195
#> hsa-let-7b-3p.iso   1144   2852   1984   1722    875    749
#> hsa-let-7b-5p.iso  32629  46634  45151  32215  19793  14587