Differential expression analysis with DESeq2

This function does differential expression analysis with DESeq2::DESeq2-package using the specific formula. It will return a DESeq2::DESeqDataSet object.

isoDE(ids, formula = NULL, ...)

Arguments

ids
Object of class IsomirDataSeq.

formula
Formula used for DE analysis.

...
Options to pass to isoCounts() including ref, iso5, iso3, add, subs and seed parameters.

Value

DESeq2::DESeqDataSet object. To get the differential expression isomiRs, use DESeq2::results() from DESeq2 package. This allows to ask for different contrast without calling again isoDE(). Read results manual to know how to access all the information.

Details

First, this function collapses all isomiRs in different types. Read more at isoCounts() to know the different options available to collapse isomiRs.

After that, DESeq2::DESeq2-package is used to do differential expression analysis. It uses the count matrix and design experiment stored at (counts(ids) and colData(ids)) IsomirDataSeq object to construct a DESeq2::DESeqDataSet object.

Examples

data(mirData)
ids <- isoCounts(mirData, minc=10, mins=6)
dds <- isoDE(mirData, formula=~group)

#> converting counts to integer mode