Clustering miRNAs-genes pairs in similar pattern expression

Clustering miRNAs-genes pairs

isoNetwork(mirna_rse, gene_rse, target, org, summarize = "group",
  genename = "ENSEMBL", min_cor = -0.6)

Arguments

mirna_rse
SummarizedExperiment::SummarizedExperiment with miRNA information. See details.

gene_rse
SummarizedExperiment::SummarizedExperiment with gene information. See details.

target
Matrix with miRNAs (columns) and genes (rows) target prediction (1 if it is a target, 0 if not).

org
AnnotationDbi::AnnotationDb obejct. For example:(org.Mm.eg.db).

summarize
Character column name in colData(rse) to use to group samples and compare betweem miRNA/gene expression.

genename
Character keytype of the gene names in gene_rse object.

min_cor
Numeric cutoff to consider a miRNA to regulate a target.

Value

list with network information

Details

This function will correlate miRNA and gene expression data using a specific metadata variable to group samples and detect pattern of expression that will be annotated with GO terms. mirna_rse and gene_rse can be created using the following code: mi_rse = SummarizedExperiment(assays=SimpleList(norm=mirna_matrix), colData, metadata=list(sign=mirna_keep))

where, mirna_matrix is the normalized counts expression, colData is the metadata information and mirna_keep the list of miRNAs to be used by this function.

Examples

library(org.Mm.eg.db)

#> Loading required package: AnnotationDbi

#> 
#> Attaching package: ‘AnnotationDbi’

#> The following object is masked from ‘package:isomiRs’:
#> 
#>     select

#> The following object is masked from ‘package:dplyr’:
#> 
#>     select

#> 

library(clusterProfiler)

#> Loading required package: DOSE

#> 

#> DOSE v3.3.2  For help: https://guangchuangyu.github.io/DOSE
#> 
#> If you use DOSE in published research, please cite:
#> Guangchuang Yu, Li-Gen Wang, Guang-Rong Yan, Qing-Yu He. DOSE: an R/Bioconductor package for Disease Ontology Semantic and Enrichment analysis. Bioinformatics 2015, 31(4):608-609

#> clusterProfiler v3.5.6  For help: https://guangchuangyu.github.io/clusterProfiler
#> 
#> If you use clusterProfiler in published research, please cite:
#> Guangchuang Yu., Li-Gen Wang, Yanyan Han, Qing-Yu He. clusterProfiler: an R package for comparing biological themes among gene clusters. OMICS: A Journal of Integrative Biology. 2012, 16(5):284-287.

#> 
#> Attaching package: ‘clusterProfiler’

#> The following object is masked from ‘package:purrr’:
#> 
#>     simplify

data(isoExample)
# ego <- enrichGO(row.names(assay(gene_ex_rse, "norm")),
#                 org.Mm.eg.db, "ENSEMBL", ont = "BP")
# data = isoNetwork(mirna_ex_rse, gene_ex_rse, ma_ex,
#                   org = slot(ego, "result"))
# isoPlotNet(data)