Partial Least Squares Discriminant Analysis for IsomirDataSeq

Use PLS-DA method with the normalized count data to detect the most important features (miRNAs/isomiRs) that explain better the group of samples given by the experimental design. It is a supervised clustering method with permutations to calculate the significance of the analysis.

isoPLSDA(ids, group, validation = NULL, learn = NULL, test = NULL,
 tol = 0.001, nperm = 400, refinment = FALSE, vip = 1.2)

Arguments

ids
Object of class IsomirDataSeq

group
Column name in colData(ids) to use as variable to explain.

validation
Type of validation, either NULL or "learntest". Default NULL.

learn
Optional vector of indexes for a learn-set. Only used when validation="learntest". Default NULL.

test
Optional vector of indices for a test-set. Only used when validation="learntest". Default NULL

tol
Tolerance value based on maximum change of cumulative R-squared coefficient for each additional PLS component. Default tol=0.001.

nperm
Number of permutations to compute the PLD-DA p-value based on R2 magnitude. Default nperm=400.

refinment
Logical indicating whether a refined model, based on filtering out variables with low VIP values.

vip
Variance Importance in Projection threshold value when a refinement process is considered. Default vip=1.2 .

Value

A base::list with the following elements: R2Matrix (R-squared coefficients of the PLS model), components (of the PLS, similar to PCs in a PCA), vip (most important isomiRs/miRNAs), group (classification of the samples), p.value and R2PermutationVector obtained by the permutations.

If the option refinment is set to TRUE, then the following elements will appear: R2RefinedMatrix and componentsRefinedModel (R-squared coefficients of the PLS model only using the most important miRNAs/isomiRs). As well, p.valRefined and R2RefinedPermutationVector with p-value and R2 of the permutations where samples were randomized. And finally, p.valRefinedFixed and R2RefinedFixedPermutationVector with p-value and R2 of the permutations where miRNAs/isomiRs were randomized.

Details

Partial Least Squares Discriminant Analysis (PLS-DA) is a technique specifically appropriate for analysis of high dimensionality data sets and multicollinearity (Perez-Enciso, 2013). PLS-DA is a supervised method (i.e. makes use of class labels) with the aim to provide a dimension reduction strategy in a situation where we want to relate a binary response variable (in our case young or old status) to a set of predictor variables. Dimensionality reduction procedure is based on orthogonal transformations of the original variables (miRNAs/isomiRs) into a set of linearly uncorrelated latent variables (usually termed as components) such that maximizes the separation between the different classes in the first few components (Xia, 2011). We used sum of squares captured by the model (R2) as a goodness of fit measure.

We implemented this method using the DiscriMiner::DiscriMiner-package into isoPLSDA() function. The output p-value of this function will tell about the statistical significant of the group separation using miRNA/isomiR expression data.

Read more about the parameters related to the PLS-DA directly from DiscriMiner::plsDA() function.

References

Perez-Enciso, Miguel and Tenenhaus, Michel. Prediction of clinical outcome with microarray data: a partial least squares discriminant analysis (PLS-DA) approach. Human Genetics. 2003.

Xia, Jianguo and Wishart, David S. Web-based inference of biological patterns, functions and pathways from metabolomic data using MetaboAnalyst. Nature Protocols. 2011.

Examples

data(mirData)
# Only miRNAs with > 10 reads in all samples.
ids <- isoCounts(mirData, minc=10, mins=6)
ids <- isoNorm(ids, formula=~condition)

#> converting counts to integer mode

pls.ids = isoPLSDA(ids, "condition", nperm = 2)
cat(paste0("pval:",pls.ids$p.val))

#> pval:1

cat(paste0("components:",pls.ids$components))

#> components:14.3066596178107 components:-11.769315771148 components:14.4142576599142 components:6.91586361823262 components:-5.94565855757508 components:7.83823453786082 components:6.37966219440615 components:0.77047592389204 components:-3.48189388210199 components:-2.58942324451421 components:-6.46917735610708 components:-6.4767019837019 components:-8.74767309535914 components:-5.14530966160911 components:0.0106305759548893 components:16.885752026911 components:0.973708513256021 components:-2.28606297390309 components:6.9122282769113 components:-2.7637698249964 components:-3.37470489250291 components:32.0166670455131 components:-6.71678158378238 components:-7.75123811407566 components:-12.4101001782671 components:-9.5375605666833 components:-3.59033899564165 components:-8.36842930869385 components:0.820624147266242 components:1.51484239098632 components:-3.4029141894967 components:3.67521480108773 components:4.2155788897753 components:-2.77960349911779 components:1.95739528644248 components:-1.27071646976908 components:5.24843961107633 components:4.80080523104672 components:4.88542812875792 components:0.0463860216414878 components:-7.28795535480501 components:-12.4235249948919 components:-7.29467341274792 components:-8.276493798084 components:-9.45333734444255 components:6.61283427313152 components:5.61779362821582 components:8.0411408840617 components:7.51143466051071 components:4.30709971092472 components:-6.54173107493968 components:-2.82333600009153 components:4.38827637645591 components:-3.97939889128885 components:0.313488015981479 components:1.57690297231265 components:4.48982190582283 components:0.269206441717426 components:-4.79707522440519 components:-1.68513893453653 components:2.72459729928857 components:1.11188636423195 components:-0.872068707789588 components:-1.78913372280393 components:0.52303455297383 components:8.8547056334723 components:-4.49577918183526 components:-8.47439890847455 components:1.04273069809177 components:3.09761178424635