Method to select specific miRNAs from an IsomirDataSeq object.

This method allows to select a miRNA and all its isomiRs from the count matrix.

isoSelect.IsomirDataSeq(object, mirna, minc = 10)

# S4 method for IsomirDataSeq
isoSelect(object, mirna, minc = 10)

Arguments

object
A IsomirDataSeq object.

mirna
String referring to the miRNA to show.

minc
Minimum number of isomiR reads needed to be included in the table.

Value

S4Vectors::DataFrame with count information. The row.names show the isomiR names, and each of the columns shows the counts for this isomiR in that sample. Mainly, it will return the count matrix only for isomiRs belonging to the miRNA family given by the mirna parameter. IsomiRs need to have counts bigger than minc parameter at least in one sample to be included in the output. Annotation of isomiRs follows these rules:

miRNA name

mismatches

additions

5 trimming events

3 trimming events

Examples

data(mirData)
# To select isomiRs from let-7a-5p miRNA
# and with 10000 reads or more.
isoSelect(mirData, mirna="hsa-let-7a-5p", minc=10000)

#> DataFrame with 2 rows and 14 columns
#>                              cc1       cc2       cc3       cc4       cc5
#>                        <numeric> <numeric> <numeric> <numeric> <numeric>
#> hsa-let-7a-5p             235825    171354    149541    180654    168884
#> hsa-let-7a-5p;iso_3p:t     50920     52403     35525     53213     51499
#>                              cc6       cc7       ct1       ct2       ct3
#>                        <numeric> <numeric> <numeric> <numeric> <numeric>
#> hsa-let-7a-5p             107430    153061    143030    163569    114028
#> hsa-let-7a-5p;iso_3p:t     30659     45492     42878     35795     24993
#>                              ct4       ct5       ct6       ct7
#>                        <numeric> <numeric> <numeric> <numeric>
#> hsa-let-7a-5p             123454    133092    158909    140272
#> hsa-let-7a-5p;iso_3p:t     34512     36973     47873     39585