Example of IsomirDataSeq with human brain miRNA counts data

This data set is the object return by IsomirDataSeqFromFiles. It contains miRNA count data from 14 samples: 7 control individuals (pc) and 7 patients with Parkinson's disease in early stage (Pantano et al, 2016). Use colData to see the experiment design.

data("mirData")

Format

a IsomirDataSeq class.

Source

Data is available from GEO dataset under accession number GSE97285

Every sample was analyzed with seqbuster tool, see http://seqcluster.readthedocs.org/mirna_annotation.html for more details. You can get same files running the small RNA-seq pipeline from https://github.com/chapmanb/bcbio-nextgen.

bcbio_nextgen was used for the full analysis. library(isomiRs) files = list.files(file.path(root_path),pattern = "mirbase-ready", recursive = T,full.names = T) metadata_fn = list.files(file.path(root_path), pattern = "summary.csv$",recursive = T, full.names = T) metadata = read.csv(metadata_fn, row.names="sample_id") condition = names(metadata)[1] mirData <- IsomirDataSeqFromFiles(files[rownames(design)], metadata)

References

Pantano L, Friedlander MR, Escaramis G, Lizano E et al. Specific small-RNA signatures in the amygdala at premotor and motor stages of Parkinson's disease revealed by deep sequencing analysis. Bioinformatics 2016 Mar 1;32(5):673-81. PMID: 26530722