Loads miRNA annotation from seqbuster tool or pre-processed data.

This function parses output of seqbuster tool to allow isomiRs/miRNAs analysis of samples in different groups such as characterization, differential expression and clustering. It creates an IsomirDataSeq object.

IsomirDataSeqFromFiles(
  files,
  coldata,
  rate = 0.2,
  canonicalAdd = TRUE,
  uniqueMism = TRUE,
  uniqueHits = FALSE,
  design = ~1L,
  minHits = 1L,
  header = TRUE,
  skip = 0,
  quiet = TRUE,
  ...
)

Arguments

files
files with the output of seqbuster tool

coldata
data frame containing groups for each sample

rate
minimum counts fraction to consider a mismatch a real mutation

canonicalAdd
boolean only keep A/T non-template addition. All non-template nucleotides at the 3' end will be removed if they contain C/G nts.

uniqueMism
boolean only keep mutations that have a unique hit to one miRNA molecule. For instance, if the sequence map to two different miRNAs, then it would be removed.

uniqueHits
boolean whether filtering ambigous sequences or not.

design
a formula to pass to DESeq2::DESeqDataSet

minHits
Minimum number of reads in the sample to consider it in the final matrix.

header
boolean to indicate files contain headers

skip
skip first line when reading files

quiet
boolean indicating to print messages while reading files. Default FALSE.

...
arguments provided to SummarizedExperiment and IsomirDataSeqFromRawData. including rowData.

Value

IsomirDataSeq class object.

Details

This function parses the output of http://seqcluster.readthedocs.org/mirna_annotation.html for each sample to create a count matrix for isomiRs, miRNAs or isomiRs grouped in types (i.e all sequences with variations at 5' but ignoring any other type). It creates IsomirDataSeq object (see link to example usage of this class) to allow visualization, queries, differential expression analysis and clustering. To create the IsomirDataSeq, it parses the isomiRs files, and generates an initial matrix having all isomiRs detected among samples. As well, it creates a summary for each isomiR type (trimming, addition and substitution) to visualize general isomiRs distribution.

Examples

path <- system.file("extra", package="isomiRs")
fn_list <- list.files(path, pattern="mirna", full.names = TRUE)
de <- data.frame(row.names=c("f1" , "f2"),
                 condition = c("newborn", "newborn"))
ids <- IsomirDataSeqFromFiles(fn_list, coldata=de)

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 1.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 2.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 4.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 5.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 6.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 13.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 14.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 15.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 16.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 17.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 18.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 19.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 20.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 21.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 22.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 23.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 24.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 25.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 26.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 27.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 28.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 29.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 36.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 37.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 40.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 41.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 42.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 43.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 44.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 45.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 46.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 49.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 51.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 52.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 53.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 61.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 62.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 63.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 67.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 68.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 69.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 70.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 71.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 72.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 73.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 76.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 81.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 91.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 92.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 93.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 94.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 95.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 96.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 98.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 99.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 100.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 101.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 102.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 103.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 106.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 107.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 108.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 109.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 110.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 112.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 113.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 114.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 115.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 116.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 117.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 118.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 119.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 120.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 121.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 122.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 123.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 124.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 125.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 126.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 128.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 129.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 131.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 132.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 133.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 134.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 135.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 136.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 137.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 138.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 139.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 140.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 146.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 147.

#> Total samples filtered due to low number of hits: 0

head(counts(ids))

#>                f1 f2
#> hsa-let-7a-5p   0  3
#> hsa-let-7b-5p   0 22
#> hsa-let-7d-3p  62  0
#> hsa-let-7d-5p   6  0
#> hsa-let-7e-5p   3  0
#> hsa-let-7f-5p 151  0

IsomirDataSeqFromRawData(metadata(ids)[["rawData"]], de)

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 1.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 2.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 4.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 5.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 6.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 10.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 11.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 12.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 13.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 14.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 15.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 16.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 17.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 18.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 19.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 20.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 21.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 24.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 26.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 27.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 28.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 29.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 30.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 31.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 32.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 34.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 36.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 37.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 38.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 40.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 41.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 43.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 44.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 45.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 46.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 49.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 50.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 54.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 55.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 56.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 57.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 59.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 60.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 61.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 62.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 63.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 65.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 66.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 67.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 68.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 69.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 71.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 72.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 73.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 74.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 75.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 76.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 77.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 78.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 79.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 80.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 81.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 82.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 84.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 85.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 87.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 88.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 89.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 90.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 91.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 92.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 93.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 97.

#> Warning: Problem with `mutate()` input `prop`.
#> ℹ Chi-squared approximation may be incorrect
#> ℹ Input `prop` is `list(tidy(prop.test(value, total, pctco, alternative = "greater")))`.
#> ℹ The error occurred in row 98.

#> class: IsomirDataSeq 
#> dim: 84 2 
#> metadata(2): sequences rawData
#> assays(1): counts
#> rownames(84): hsa-let-7a-5p hsa-let-7b-5p ... hsa-miR-98-5p
#>   hsa-miR-99b-5p
#> rowData names(0):
#> colnames(2): f1 f2
#> colData names(1): condition