IsomirDataSeqFromFiles loads miRNA annotation from seqbuster tool

This function parses output of seqbuster tool to allow isomiRs/miRNAs analysis of samples in different groups such as characterization, differential expression and clustering. It creates an IsomirDataSeq object.

IsomirDataSeqFromFiles(files, coldata, rate = 0.2, canonicalAdd = TRUE,
  uniqueMism = TRUE, design = ~1, header = TRUE, skip = 0,
  quiet = TRUE, ...)

Arguments

files
files with the output of seqbuster tool

coldata
data frame containing groups for each sample

rate
minimum counts fraction to consider a mismatch a real mutation

canonicalAdd
boolean only keep A/T non-template addition. All non-template nucleotides at the 3' end will be removed if they contain C/G nts.

uniqueMism
boolean only keep mutations that have a unique hit to one miRNA molecule

design
a formula to pass to DESeqDataSet

header
boolean to indicate files contain headers

skip
skip first line when reading files

quiet
boolean indicating to print messages while reading files. Default FALSE.

...
arguments provided to SummarizedExperiment including rowData.

Value

IsomirDataSeq class object.

Details

This function parses the output of http://seqcluster.readthedocs.org/mirna_annotation.html for each sample to create a count matrix for isomiRs, miRNAs or isomiRs grouped in types (i.e all sequences with variations at 5' but ignoring any other type). It creates IsomirDataSeq object (see link to example usage of this class) to allow visualization, queries, differential expression analysis and clustering. To create the IsomirDataSeq, it parses the isomiRs files, and generates an initial matrix having all isomiRs detected among samples. As well, it creates a summary for each isomiR type (trimming, addition and substitution) to visualize general isomiRs distribution.

Examples

path <- system.file("extra", package="isomiRs")
fn_list <- list.files(path, full.names = TRUE)
de <- data.frame(row.names=c("f1" , "f2"), condition = c("newborn", "newborn"))
ids <- IsomirDataSeqFromFiles(fn_list, coldata=de)

head(counts(ids))

#>               f1 f2
#> hsa-let-7a-5p  0 18
#> hsa-let-7b-5p  0  6
#> hsa-let-7c-5p  0 61
#> hsa-let-7f-5p  8  7
#> hsa-let-7g-5p 74  3
#> hsa-let-7i-5p  0  2