Accessors for the count matrix of a IsomirDataSeq object.

The counts slot holds the count data as a matrix of non-negative integer count values, one row for each isomiR, and one column for each sample. The normalized matrix can be obtained by using the parameter norm=TRUE.

counts.IsomirDataSeq(object, norm = FALSE)

# S4 method for IsomirDataSeq
counts(object, norm = FALSE)

# S4 method for IsomirDataSeq,matrix
counts(object) <- value

Arguments

object
A IsomirDataSeq object.

norm
Boolean, return log2-normalized counts.

value
An integer matrix.

Value

base::matrix with raw or normalized count data.

Author

Lorena Pantano

Examples

data(mirData)
head(counts(mirData))

#>                    cc1    cc2    cc3    cc4    cc5    cc6    cc7    ct1    ct2
#> hsa-let-7a-2-3p      5     13      4      9     11      6      7      0      4
#> hsa-let-7a-3p      767    707    630    609    731    389    681    258    496
#> hsa-let-7a-5p   317730 244832 203888 260931 244784 153049 220563 208511 222066
#> hsa-let-7b-3p     1037   1949    679   1385   1884    697   1499    338    931
#> hsa-let-7b-5p    69671  64702  42347  65322  60767  34975  56875  22215  45069
#> hsa-let-7c-3p       36     16     25     40     26     23     35      3     30
#>                    ct3    ct4    ct5    ct6    ct7
#> hsa-let-7a-2-3p      2     12     12      8      3
#> hsa-let-7a-3p      343    635    622    452    519
#> hsa-let-7a-5p   154246 175523 189733 230690 199923
#> hsa-let-7b-3p      494   1149   1194   1431    988
#> hsa-let-7b-5p    28312  41214  43058  61258  46600
#> hsa-let-7c-3p       28     26     24     22     17