Accessors for the count matrix of a IsomirDataSeq object.

The counts slot holds the count data as a matrix of non-negative integer count values, one row for each isomiR, and one column for each sample. The normalized matrix can be obtained by using the parameter norm=TRUE.

counts.IsomirDataSeq(object, norm = FALSE)

# S4 method for IsomirDataSeq
counts(object, norm = FALSE)

# S4 method for IsomirDataSeq,matrix
counts(object) <- value

Arguments

object
A IsomirDataSeq object.

norm
Boolean, return log2-normalized counts.

value
An integer matrix.

Value

base::matrix with raw or normalized count data.

Examples

data(mirData)
head(counts(mirData))

#>                    cc1    cc2    cc3    ct2    ct3    ct4
#> hsa-let-7a-2-3p      5     13      4      4      2     12
#> hsa-let-7a-3p      767    707    630    496    343    635
#> hsa-let-7a-5p   317730 244832 203888 222066 154246 175523
#> hsa-let-7b-3p     1037   1949    679    931    494   1149
#> hsa-let-7b-5p    69671  64702  42347  45071  28312  41214
#> hsa-let-7c-3p       36     16     25     30     28     26