Annotate the rawData of the IsomirDataSeq object

Get the sequence and the name information for each isomiR, and the importance value (isomir_reads/mirna_reads) for each sample.

isoAnnotate(ids)

Arguments

ids
Object of class IsomirDataSeq.

Value

data.frame with the sequence, isomir name, and importance for each sample and isomiR.

Details

edit_mature_position represents the position at the mature sequence + nucleotide at reference + nucleotide at isomiR.

Examples

data(mirData)
head(isoAnnotate(mirData))

#>                       seq
#> 1 AAAAACCGTAGTTACTTTTGCAT
#> 2    AAAAACTGAGACTACTTTTG
#> 3   AAAAACTGAGACTACTTTTGC
#> 4  AAAAACTGAGACTACTTTTGCA
#> 5 AAAAACTGAGACTACTTTTGCAA
#> 6   AAAAACTGCAGTTACTTTTGC
#>                                                        uid    cc1      cc2
#> 1 hsa-miR-548bb-3p;iso_snp:8GA;iso_add:T;iso_5p:c;iso_3p:A     NA       NA
#> 2                                hsa-miR-548e-3p;iso_3p:ca  9.375 10.71429
#> 3                                 hsa-miR-548e-3p;iso_3p:a 34.375 10.71429
#> 4                                      hsa-miR-548e-3p;ref 21.875 64.28571
#> 5                                hsa-miR-548e-3p;iso_add:A 34.375       NA
#> 6                                hsa-miR-548ah-3p;iso_5p:c 15.000 11.76471
#>        cc3        cc4 cc5      cc6 cc7      ct1      ct2       ct3 ct4
#> 1       NA 100.000000  NA       NA  NA       NA       NA        NA  NA
#> 2 15.15152   6.896552  NA       NA  NA 19.04762       NA  8.333333  20
#> 3 30.30303  41.379310  28 66.66667  25 19.04762 47.61905 29.166667  20
#> 4 27.27273  17.241379  44 33.33333  60 38.09524 23.80952 16.666667  30
#> 5 21.21212  13.793103  12       NA  15 23.80952 19.04762 25.000000  30
#> 6 23.07692         NA  NA       NA  NA       NA       NA        NA  NA
#>         ct5      ct6      ct7 edit_mature_position
#> 1        NA       NA       NA                 9:AG
#> 2  8.695652       NA       NA                   NA
#> 3 30.434783 13.04348 46.15385                   NA
#> 4 30.434783 34.78261 30.76923                   NA
#> 5 17.391304 43.47826 23.07692                   NA
#> 6        NA       NA       NA                   NA