Create count matrix with different summarizing options

This function collapses isomiRs into different groups. It is a similar concept than how to work with gene isoforms. With this function, different changes can be put together into a single miRNA variant. For instance all sequences with variants at 3' end can be considered as different elements in the table or analysis having the following naming hsa-miR-124a-5p.iso.t3:AAA.

isoCounts(ids, ref = FALSE, iso5 = FALSE, iso3 = FALSE, add = FALSE,
  subs = FALSE, seed = FALSE, minc = 1, mins = 1)

Arguments

ids
Object of class IsomirDataSeq.

ref
Differentiate reference miRNA from rest.

iso5
Differentiate trimming at 5 miRNA from rest.

iso3
Differentiate trimming at 3 miRNA from rest.

add
Differentiate additions miRNA from rest.

subs
Differentiate nt substitution miRNA from rest.

seed
Differentiate changes in 2-7 nts from rest.

minc
Int minimum number of isomiR sequences to be included.

mins
Int minimum number of samples with number of sequences bigger than minc counts.

Value

IsomirDataSeq object with new count table. The count matrix can be access with counts(ids).

Details

You can merge all isomiRs into miRNAs by calling the function only with the first parameter isoCounts(ids). You can get a table with isomiRs altogether and the reference miRBase sequences by calling the function with ref=TRUE. You can get a table with 5' trimming isomiRS, miRBase reference and the rest by calling with isoCounts(ids, ref=TRUE, iso5=TRUE). If you set up all parameters to TRUE, you will get a table for each different sequence mapping to a miRNA (i.e. all isomiRs).

Examples for the naming used for the isomiRs are at http://seqcluster.readthedocs.org/mirna_annotation.html#mirna-annotation.

Examples

data(mirData)
ids <- isoCounts(mirData, ref=TRUE)
head(counts(ids))

#>                        cc1    cc2    cc3    ct2    ct3    ct4
#> hsa-let-7a-2-3p.iso      2      7      2      0      2      6
#> hsa-let-7a-2-3p.ref      3      6      2      4      0      6
#> hsa-let-7a-3p.iso      653    593    543    448    312    553
#> hsa-let-7a-3p.ref      114    114     87     48     31     82
#> hsa-let-7a-5p.iso    81905  73478  54347  58497  40218  52069
#> hsa-let-7a-5p.ref   235825 171354 149541 163569 114028 123454

# taking into account isomiRs and reference sequence.
ids <- isoCounts(mirData, ref=TRUE, minc=10, mins=6)
head(counts(ids))

#>                      cc1    cc2    cc3    ct2    ct3    ct4
#> hsa-let-7a-3p.iso    653    593    543    448    312    553
#> hsa-let-7a-3p.ref    114    114     87     48     31     82
#> hsa-let-7a-5p.iso  81905  73478  54347  58497  40218  52069
#> hsa-let-7a-5p.ref 235825 171354 149541 163569 114028 123454
#> hsa-let-7b-3p.iso   1033   1729    668    808    483   1058
#> hsa-let-7b-5p.iso  31209  36938  18988  22156  14187  20585