Partial Least Squares Discriminant Analysis for IsomirDataSeq

Use PLS-DA method with the normalized count data to detect the most important features (miRNAs/isomiRs) that explain better the group of samples given by the experimental design. It is a supervised clustering method with permutations to calculate the significance of the analysis.

isoPLSDA(ids, group, validation = NULL, learn = NULL, test = NULL,
 tol = 0.001, nperm = 400, refinment = FALSE, vip = 1.2)

Arguments

ids
Object of class IsomirDataSeq

group
Column name in colData(ids) to use as variable to explain.

validation
Type of validation, either NULL or "learntest". Default NULL.

learn
Optional vector of indexes for a learn-set. Only used when validation="learntest". Default NULL.

test
Optional vector of indices for a test-set. Only used when validation="learntest". Default NULL

tol
Tolerance value based on maximum change of cumulative R-squared coefficient for each additional PLS component. Default tol=0.001.

nperm
Number of permutations to compute the PLD-DA p-value based on R2 magnitude. Default nperm=400.

refinment
Logical indicating whether a refined model, based on filtering out variables with low VIP values.

vip
Variance Importance in Projection threshold value when a refinement process is considered. Default vip=1.2 .

Value

A base::list with the following elements: R2Matrix (R-squared coefficients of the PLS model), components (of the PLS, similar to PCs in a PCA), vip (most important isomiRs/miRNAs), group (classification of the samples), p.value and R2PermutationVector obtained by the permutations.

If the option refinment is set to TRUE, then the following elements will appear: R2RefinedMatrix and componentsRefinedModel (R-squared coefficients of the PLS model only using the most important miRNAs/isomiRs). As well, p.valRefined and R2RefinedPermutationVector with p-value and R2 of the permutations where samples were randomized. And finally, p.valRefinedFixed and R2RefinedFixedPermutationVector with p-value and R2 of the permutations where miRNAs/isomiRs were randomized.

Details

Partial Least Squares Discriminant Analysis (PLS-DA) is a technique specifically appropriate for analysis of high dimensionality data sets and multicollinearity (Perez-Enciso, 2013). PLS-DA is a supervised method (i.e. makes use of class labels) with the aim to provide a dimension reduction strategy in a situation where we want to relate a binary response variable (in our case young or old status) to a set of predictor variables. Dimensionality reduction procedure is based on orthogonal transformations of the original variables (miRNAs/isomiRs) into a set of linearly uncorrelated latent variables (usually termed as components) such that maximizes the separation between the different classes in the first few components (Xia, 2011). We used sum of squares captured by the model (R2) as a goodness of fit measure.

We implemented this method using the DiscriMiner::DiscriMiner-package into isoPLSDA() function. The output p-value of this function will tell about the statistical significant of the group separation using miRNA/isomiR expression data.

Read more about the parameters related to the PLS-DA directly from DiscriMiner::plsDA() function.

References

Perez-Enciso, Miguel and Tenenhaus, Michel. Prediction of clinical outcome with microarray data: a partial least squares discriminant analysis (PLS-DA) approach. Human Genetics. 2003.

Xia, Jianguo and Wishart, David S. Web-based inference of biological patterns, functions and pathways from metabolomic data using MetaboAnalyst. Nature Protocols. 2011.

Examples

data(mirData)
# Only miRNAs with > 10 reads in all samples.
ids <- isoCounts(mirData, minc=10, mins=6)
ids <- isoNorm(ids, formula=~condition)

#> converting counts to integer mode

pls.ids = isoPLSDA(ids, "condition", nperm = 2)
cat(paste0("pval:",pls.ids$p.val))

#> pval:1

cat(paste0("components:",pls.ids$components))

#> components:14.3066366361985 components:-11.76927567062 components:14.4142252427394 components:6.9158409610585 components:-5.94561492551998 components:7.83823308619104 components:6.37965375808312 components:0.770507005119016 components:-3.48191038598088 components:-2.58943669983855 components:-6.46919365801227 components:-6.47670533056536 components:-8.7476621770196 components:-5.14529784183292 components:0.0105721634718636 components:16.8857951387221 components:0.973658919092283 components:-2.28604793052447 components:6.91225644808857 components:-2.7637813863646 components:-3.37469851222244 components:32.0166888703228 components:-6.71679573377505 components:-7.75125734144394 components:-12.4100891867135 components:-9.53754160723886 components:-3.59032659166325 components:-8.3684332497515 components:0.820586539895829 components:1.51482587885364 components:-3.40290150429687 components:3.67527063642309 components:4.21555882608403 components:-2.77960362499496 components:1.95740375627186 components:-1.27070905183756 components:5.24841830967773 components:4.80078945240899 components:4.8854437350528 components:0.0463901068743711 components:-7.28792845151067 components:-12.4235446089023 components:-7.29470420845995 components:-8.27652061849203 components:-9.45331298875878 components:6.61286418208011 components:5.61781625740843 components:8.04113181774385 components:7.51141145080473 components:4.30707971172336 components:-6.54176560987081 components:-2.82335019627297 components:4.38827706025239 components:-3.97939488805987 components:0.313543609283786 components:1.57692442061777 components:4.48980054184539 components:0.269204174738415 components:-4.79706093257499 components:-1.68513498652605 components:2.72461776639589 components:1.11188072353706 components:-0.872083818411135 components:-1.78912107548238 components:0.523022027610943 components:8.85472938196647 components:-4.49574088568728 components:-8.47441129109108 components:1.04271279247261 components:3.09758558120614