Method to select specific miRNAs from an IsomirDataSeq object.

This method allows to select a miRNA and all its isomiRs from the count matrix.

isoSelect.IsomirDataSeq(object, mirna, minc = 10)

# S4 method for IsomirDataSeq
isoSelect(object, mirna, minc = 10)

Arguments

object
A IsomirDataSeq object.

mirna
String referring to the miRNA to show.

minc
Minimum number of isomiR reads needed to be included in the table.

Value

S4Vectors::DataFrame with count information. The row.names show the isomiR names, and each of the columns shows the counts for this isomiR in that sample. Mainly, it will return the count matrix only for isomiRs belonging to the miRNA family given by the mirna parameter. IsomiRs need to have counts bigger than minc parameter at least in one sample to be included in the output. Annotation of isomiRs follows these rules:

miRNA name

mismatches

additions

5 trimming events

3 trimming events

Examples

data(mirData)
# To select isomiRs from let-7a-5p miRNA
# and with 10000 reads or more.
isoSelect(mirData, mirna="hsa-let-7a-5p", minc=10000)

#> DataFrame with 2 rows and 7 columns
#>                                               id       cc1       cc2       cc3
#>                                      <character> <numeric> <numeric> <numeric>
#> 1 hsa-let-7a-5p 0 0 0 0 : TGAGGTAGTAGGTTGTATAGTT    235825    171354    149541
#> 2  hsa-let-7a-5p 0 0 0 t : TGAGGTAGTAGGTTGTATAGT     50920     52403     35525
#>         ct2       ct3       ct4
#>   <numeric> <numeric> <numeric>
#> 1    163569    114028    123454
#> 2     35795     24993     34512